+++ sdk/containerservice/arm-containerservice/CHANGELOG.md
@@ -1,24 +1,32 @@
 # Release History
     
-## 19.1.0-beta.1 (2023-06-12)
+## 19.1.0-beta.2 (2023-06-21)
     
 **Features**
 
   - Added operation group ManagedClusterSnapshots
   - Added operation group TrustedAccessRoleBindings
   - Added operation group TrustedAccessRoles
+  - Added operation ManagedClusters.getMeshRevisionProfile
+  - Added operation ManagedClusters.getMeshUpgradeProfile
+  - Added operation ManagedClusters.listMeshRevisionProfiles
+  - Added operation ManagedClusters.listMeshUpgradeProfiles
   - Added Interface AbsoluteMonthlySchedule
   - Added Interface AgentPoolNetworkProfile
   - Added Interface AgentPoolWindowsProfile
   - Added Interface ClusterUpgradeSettings
+  - Added Interface CompatibleVersions
   - Added Interface ContainerServiceNetworkProfileKubeProxyConfig
   - Added Interface ContainerServiceNetworkProfileKubeProxyConfigIpvsConfig
   - Added Interface DailySchedule
   - Added Interface DateSpan
   - Added Interface GuardrailsProfile
   - Added Interface IPTag
+  - Added Interface IstioCertificateAuthority
   - Added Interface IstioComponents
+  - Added Interface IstioEgressGateway
   - Added Interface IstioIngressGateway
+  - Added Interface IstioPluginCertificateAuthority
   - Added Interface IstioServiceMesh
   - Added Interface MaintenanceWindow
   - Added Interface ManagedClusterIngressProfile
@@ -26,6 +34,12 @@
   - Added Interface ManagedClusterNodeResourceGroupProfile
   - Added Interface ManagedClusterPropertiesForSnapshot
   - Added Interface ManagedClusterSecurityProfileNodeRestriction
+  - Added Interface ManagedClustersGetMeshRevisionProfileOptionalParams
+  - Added Interface ManagedClustersGetMeshUpgradeProfileOptionalParams
+  - Added Interface ManagedClustersListMeshRevisionProfilesNextOptionalParams
+  - Added Interface ManagedClustersListMeshRevisionProfilesOptionalParams
+  - Added Interface ManagedClustersListMeshUpgradeProfilesNextOptionalParams
+  - Added Interface ManagedClustersListMeshUpgradeProfilesOptionalParams
   - Added Interface ManagedClusterSnapshot
   - Added Interface ManagedClusterSnapshotListResult
   - Added Interface ManagedClusterSnapshotsCreateOrUpdateOptionalParams
@@ -37,9 +51,17 @@
   - Added Interface ManagedClusterSnapshotsListOptionalParams
   - Added Interface ManagedClusterSnapshotsUpdateTagsOptionalParams
   - Added Interface ManagedClusterWorkloadAutoScalerProfileVerticalPodAutoscaler
+  - Added Interface MeshRevision
+  - Added Interface MeshRevisionProfile
+  - Added Interface MeshRevisionProfileList
+  - Added Interface MeshRevisionProfileProperties
+  - Added Interface MeshUpgradeProfile
+  - Added Interface MeshUpgradeProfileList
+  - Added Interface MeshUpgradeProfileProperties
   - Added Interface NetworkMonitoring
   - Added Interface NetworkProfileForSnapshot
   - Added Interface PortRange
+  - Added Interface ProxyResource
   - Added Interface RelativeMonthlySchedule
   - Added Interface Schedule
   - Added Interface ServiceMeshProfile
@@ -63,6 +85,12 @@
   - Added Type Alias IpvsScheduler
   - Added Type Alias IstioIngressGatewayMode
   - Added Type Alias Level
+  - Added Type Alias ManagedClustersGetMeshRevisionProfileResponse
+  - Added Type Alias ManagedClustersGetMeshUpgradeProfileResponse
+  - Added Type Alias ManagedClustersListMeshRevisionProfilesNextResponse
+  - Added Type Alias ManagedClustersListMeshRevisionProfilesResponse
+  - Added Type Alias ManagedClustersListMeshUpgradeProfilesNextResponse
+  - Added Type Alias ManagedClustersListMeshUpgradeProfilesResponse
   - Added Type Alias ManagedClusterSnapshotsCreateOrUpdateResponse
   - Added Type Alias ManagedClusterSnapshotsGetResponse
   - Added Type Alias ManagedClusterSnapshotsListByResourceGroupNextResponse